Raise cake404 if file is not .css nor .js
When working with third party css files, background images
maybe specified in the current directory.

Eg:
body {
	background: url('bg.png');
}

This causes requests for 'bg.png' to be passed over to asset_compress,
causing it to generate invalid bg.png file in the cacheFilePath directory.

This patch makes asset_compress abort processing and returns a 404 error
to browser.

Signed-off-by: Mark Story <mark@mark-story.com>