Remove dynamic build file support.
While sometimes useful it also adds a pile of complexity to the plugin
and I think it is of very questionable value. Having assets in an ini
file makes it much easier to reason about what assets are available in
an application/plugin. Having centralized configuration also makes other
tooling simpler.