Add code and tests for new FilterRegistry.
The Factory class will create a filter registry based on the config
data. The filter registry can then be used to fetch filters for a given
asset as necessary.  The new filter collection doesn't know how to build
filters only use them. The construction logic will be placed in the
Factory.