Issue #329 -- Editing a SASS partial in v0.20 does not trigger the output CSS file to be regenerated when "always enable controller" is false
Alter the `isFresh()` method to detect a SASS or LESS file being processed. If one of those types, get the file contents and check for any import statements. If any import statements, check the freshness of those files.

Kept it at one level of import checking. Not sure how common nesting imports is. I'm not familiar enough with LESS or have an environment to test that code path, but the SASS path works in my local environment.